==> 2-Integration_STACAS.Rmd <==
---
title: "CD8 TIL integration with STACAS"
author: "Paul Gueguen, Massimo Andreatta, and Santiago  Carmona"
date: "`r format(Sys.Date(),'%e de %B, %Y')`"
output: 
  rmdformats::downcute:
    lightbox: true
    thumbnails: false
    self_contained: true
    gallery: true
    code_folding: show
  pkgdown:
    as_is: true
---

```{r, include=FALSE, fig.width=16, fig.height=12}
library(renv)
renv::restore()
library(Seurat)
library(ggplot2)
library(scGate)
library(STACAS)
library(SignatuR)
library(dplyr)
#options(future.globals.maxSize= 5000*1024^2)
```

We start from object of high-quality cells and balanced datasets generated by 1-QC_CD8_atlas.Rmd.

# Set up parameters

```{r}
path_output <- "./out/"
npcs <- 50
nfeatures <- 800
seed <- 123
set.seed(seed)
umap.neighbors <- 30
semisup=TRUE

# Setup color palette
mycols <- c('NaiveLike' = '#b7d2e0', 'CM' = '#da6f6f','EM'= '#72b28a','TEMRA' = '#e5bfaf', 'TPEX' = '#aca6e0' , 'TEX' ='#f5d39f', "MAIT" = '#fdbfd4')

mycols_scGate <- c('CD8_N' = '#b7d2e0', 'CD8_EM'= '#72b28a','CD8_TEMRA' = '#e5bfaf', 'CD8_TPEX' = '#aca6e0' , 'CD8_TEX' ='#f5d39f', "CD8_MAIT" = '#fdbfd4')
```

# Load data

```{r}
seurat.list <- readRDS('cache/Seurat.list.Utility.v0.4_step9.rds')

# Select top N biggest datasets
n <- 20
ncells <- unlist(lapply(seurat.list, ncol))
sorted <- sort(ncells, decreasing = T)
which.sets <- names(sorted[1:n])
seurat.list <- seurat.list[which.sets]
```


# Simple merge of the data

Without any integration, what is the extent of batch effects? how are batch/label metrics?
```{r fig.height=3}
library(scIntegrationMetrics)

merged <- Reduce(f=merge, x=seurat.list)

merged <- FindVariableFeatures(merged, nfeatures = nfeatures) |> ScaleData() |>
  RunPCA(npcs=npcs) |> RunUMAP(reduction = "pca", dims = 1:npcs, seed.use=seed, n.neighbors = umap.neighbors)

a <- DimPlot(merged, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(merged, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_preintegration.pdf", height=4, width=9)

#Save object to evaluate metrics
saveRDS(merged, file="cache/CD8T_unintegrated.rds")
```

Also export in h5ad format, to be read by scipy
```{r}
library(SeuratDisk)
to.conv <- DietSeurat(merged, counts=T, scale.data = F)
#to.conv$scGate_multi[is.na(to.conv$scGate_multi)] <- "unknown"
SaveH5Seurat(to.conv, filename = "cache/merged.h5Seurat", overwrite = T)
Convert("cache/merged.h5Seurat", dest = "h5ad", overwrite = T)
```


Gene categories to exclude from variable genes
```{r}
bl <- c(GetSignature(SignatuR$Hs$Blocklists$Pseudogenes),
        GetSignature(SignatuR$Hs$Blocklists$HSP),
        GetSignature(SignatuR$Hs$Blocklists$`Non-coding`),
        GetSignature(SignatuR$Hs$Programs$cellCycle.G1S),
        GetSignature(SignatuR$Hs$Programs$cellCycle.G2M),
        GetSignature(SignatuR$Hs$Programs$IFN),
        GetSignature(SignatuR$Hs$Compartments$Mito),
        GetSignature(SignatuR$Hs$Compartments$Ribo),
        GetSignature(SignatuR$Hs$Compartments$TCR),
        GetSignature(SignatuR$Hs$Compartments$Immunoglobulins))
bl <- unique(bl)
bl <- list("Blocklist"=bl)

```

# Unsupervised STACAS integration

```{r}
for (i in 1:length(seurat.list)) {
  seurat.list[[i]] <- NormalizeData(seurat.list[[i]])
  seurat.list[[i]] <- FindVariableFeatures.STACAS(seurat.list[[i]],
                                                  nfeat=nfeatures,
                                                  genesBlockList = bl)
}
hvg <- SelectIntegrationFeatures(seurat.list, nfeatures = nfeatures)

stacas_anchors <- FindAnchors.STACAS(seurat.list,
                                     genesBlockList = bl,
                                     anchor.features = hvg,
                                     max.seed.objects = 20,
                                     dims = 1:npcs)

tree <- SampleTree.STACAS(stacas_anchors, obj.names = names(seurat.list),
                          hclust.method = 'ward.D2')

integrated.us <- IntegrateData.STACAS(anchorset=stacas_anchors, dims=1:npcs,
                                   sample.tree=tree)

```

See low-dim reductions
```{r}
integrated.us <- ScaleData(integrated.us) |> RunPCA(dims = 1:npcs) |>
      RunUMAP(reduction = "pca", dims = 1:npcs, seed.use=seed, n.neighbors = umap.neighbors,
              metric = "cosine")

DimPlot(integrated.us, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", label = T) +
  ggtitle('STACAS - scGate_multi') + theme(aspect.ratio = 1)

DimPlot(integrated.us, reduction = "umap",
        group.by = "SampleLabel") + theme(aspect.ratio = 1)

integrated.us$prelab <- integrated.us$scGate_multi
```
Re-run scGate to check possible non-T cells
```{r}
library(scGate)
#models <- scGate::get_scGateDB(branch = "dev", force_update = T)
models <- scGate::get_scGateDB(version = 'v0.12') #fix version for reproducibility
cd8.model <- models$human$generic$CD8TIL

integrated.us <- scGate(integrated.us, model = cd8.model, reduction = 'pca', ncores=4)
DimPlot(integrated.us)

cells.keep <- colnames(integrated.us)[integrated.us$is.pure == "Pure"]

# Remove non-pure T cells
integrated.us <- subset(integrated.us, cells = cells.keep)
# also from individual objects
seurat.list <- lapply(seurat.list, function(x) {
  subset(x, cells = colnames(x)[colnames(x) %in% cells.keep])
})

```

Save object to evaluate metrics
```{r}
integrated.us$scGate_multi <- integrated.us$prelab

saveRDS(integrated.us, file="cache/CD8T_integrated_unsup.rds")
```

# Semi-supervised STACAS integration

We will use scGate labels to guide STACAS integration.

```{r}
for (i in 1:length(seurat.list)) {
  seurat.list[[i]] <- NormalizeData(seurat.list[[i]])
  seurat.list[[i]] <- FindVariableFeatures.STACAS(seurat.list[[i]],
                                                  nfeat=nfeatures,
                                                  genesBlockList = bl)
}
hvg <- SelectIntegrationFeatures(seurat.list, nfeatures = nfeatures)

stacas_anchors <- FindAnchors.STACAS(seurat.list,
                                     cell.labels = "scGate_multi",
                                     genesBlockList = bl,
                                     anchor.features = hvg,
                                     max.seed.objects = 20,
                                     dims = 1:npcs)

tree <- SampleTree.STACAS(stacas_anchors, obj.names = names(seurat.list),
                          semisupervised = semisup, hclust.method = 'ward.D2'
                          )

integrated <- IntegrateData.STACAS(anchorset=stacas_anchors, dims=1:npcs,
                                   sample.tree=tree,
                                   semisupervised = semisup)

#Or use the wrapper
#integrated <- Run.STACAS(obj4int, anchor.features = 1000, cell.labels = "scGate_multi")
```

Reduced dimensionalities
```{r}
integrated <- ScaleData(integrated) |> RunPCA(dims = 1:npcs) |>
      RunUMAP(reduction = "pca", dims = 1:npcs, seed.use=seed, n.neighbors = umap.neighbors,
              metric = "cosine")
```

See results
```{r}
a <- DimPlot(integrated, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('ssSTACAS - scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(integrated, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_1st.pdf", height=4, width=9)


```


Save object to evaluate metrics
```{r}
saveRDS(integrated, file="cache/CD8T_integrated_1st.rds")
```


```{r fig.height=8, fig.width=13}
DefaultAssay(integrated) <- "RNA"
FeaturePlot(integrated, features=c("PDCD1","HAVCR2","TCF7","TRAV1-2","SLC4A10","KLRB1"), ncol=3)
FeaturePlot(integrated, features=c("CCR7","CRTAM","TOX","IL7R","GZMB","GZMK"), ncol=3)
FeaturePlot(integrated, features=c("XCL1","PRF1","LAG3","CTLA4","LMNA","GZMA"), ncol=3)
FeaturePlot(integrated, features=c("TRDC","TRDV1","TRDV2","TRGC1","TRGC2","TRAC"), ncol=3)
DefaultAssay(integrated) <- "integrated"
```

```{r eval=T}
integrated <- FindNeighbors(integrated, dims = 1:npcs)
integrated <- FindClusters(integrated, resolution = 0.7)

DimPlot(integrated, reduction = "umap", group.by = "seurat_clusters", label = T) + theme(aspect.ratio = 1)
table(integrated$scGate_multi, integrated$seurat_clusters)

table(integrated$SampleLabel, integrated$seurat_clusters)
```

```{r fig.height=10}
# Heatmap
markers <- FindAllMarkers(integrated, only.pos = TRUE, assay = "RNA", min.pct = 0.2, min.diff.pct = 0.1)
top15 <- markers %>% group_by(cluster) %>% top_n(n = 15, wt = avg_log2FC)

Average <- AverageExpression(integrated, return.seurat = T, assay = "RNA")
Average <- ScaleData(Average, features = rownames(markers))

DoHeatmap(Average, features = top15$gene, draw.lines = F) + scale_fill_viridis_c() + theme(text = element_text(size = 9))
ggsave("plots/HeatmapDEGs.png", height=12, width=6)
```

```{r}
for (c in unique(markers$cluster)) {
  sub <- markers[markers$cluster==c,]
  print(sub)
}
```

```{r fig.height=7}
# Vlnplot
DefaultAssay(integrated) <- "RNA"
Idents(integrated) <- integrated$seurat_clusters
genes <- c('SELL','LEF1', "TCF7", "CCR7",'IL7R',"GZMK","CX3CR1",
           "KLRG1","FGFBP2",'FCGR3A','XCL1',"XCL2",'ITGAE',
           "CRTAM","CD200",'PDCD1', "HAVCR2","GZMB","PRF1", "TOX",
           'KLRB1','TRAV1-2','SLC4A10')
VlnPlot(integrated, pt.size = 0.1, features = genes,
        stack = TRUE, flip = TRUE, assay = "RNA", fill.by = "ident")
DefaultAssay(integrated) <- "integrated"
ggsave("plots/CD8_ref_violins_clusters.png", height=12, width=6)
```


Assign provisional annotations
```{r}
table(integrated$scGate_multi, integrated$seurat_clusters, useNA = "ifany")

integrated$functional.cluster <- NA
integrated$functional.cluster[integrated$seurat_clusters %in% c(0)] <- "EM"
integrated$functional.cluster[integrated$seurat_clusters %in% c(1)] <- "CM"
integrated$functional.cluster[integrated$seurat_clusters %in% c(2)] <- NA
integrated$functional.cluster[integrated$seurat_clusters %in% c(3)] <- "TEX"
integrated$functional.cluster[integrated$seurat_clusters %in% c(4)] <- "NaiveLike"
integrated$functional.cluster[integrated$seurat_clusters %in% c(5)] <- "TPEX"
integrated$functional.cluster[integrated$seurat_clusters %in% c(6)] <- "TEMRA"
integrated$functional.cluster[integrated$seurat_clusters %in% c(7)] <- "MAIT"
integrated$functional.cluster[integrated$seurat_clusters %in% c(8)] <- NA

```
Expert annotation labels
```{r}
a <- DimPlot(integrated, reduction = "umap",
        cols = mycols, group.by = "functional.cluster", raster = T) +
  ggtitle('Exp annotation') + theme(aspect.ratio = 1)

b <- DimPlot(integrated, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_1st_wlabels.pdf", height=4, width=9)

```

Remove cells that cannot be labeled
```{r}
to.keep <- colnames(integrated)[!is.na(integrated$functional.cluster)]

integrated <- subset(integrated, cells=to.keep)
```

Re-do integration based on these labels
```{r}
DefaultAssay(integrated) <- "RNA"
seurat.list <- SplitObject(integrated, split.by = "SampleLabel")
for (i in 1:length(seurat.list)) {
  seurat.list[[i]] <- FindVariableFeatures.STACAS(seurat.list[[i]],
                                                  nfeat=nfeatures,
                                                  genesBlockList = bl)
}
hvg <- SelectIntegrationFeatures(seurat.list, nfeatures = nfeatures)

stacas_anchors <- FindAnchors.STACAS(seurat.list,
                                     cell.labels = "functional.cluster",
                                     genesBlockList = bl,
                                     anchor.features = hvg,
                                     max.seed.objects = 20,
                                     dims = 1:npcs)

tree <- SampleTree.STACAS(stacas_anchors, obj.names = names(seurat.list),
                          semisupervised = semisup,
                          hclust.method = 'ward.D2')

integrated2 <- IntegrateData.STACAS(anchorset=stacas_anchors, dims=1:npcs,
                                   sample.tree=tree,
                                   semisupervised = semisup)


integrated2 <- ScaleData(integrated2) |> RunPCA(dims = 1:npcs) |>
      RunUMAP(reduction = "pca", dims = 1:npcs, seed.use=seed, n.neighbors = umap.neighbors,
              metric = "cosine")
```

# See results
```{r}
a <- DimPlot(integrated2, reduction = "umap",
        cols = mycols, group.by = "functional.cluster", raster = T) +
  ggtitle('Exp annotation') + theme(aspect.ratio = 1)

b <- DimPlot(integrated2, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a
```

Save object to evaluate metrics
```{r}
saveRDS(integrated2, file="cache/CD8T_integrated_2nd.rds")
```

Re-cluster?
```{r}
integrated2 <- FindNeighbors(integrated2, dims = 1:npcs)
integrated2 <- FindClusters(integrated2, resolution = 0.5)

DimPlot(integrated2, reduction = "umap", group.by = "seurat_clusters", label = T) 
table(integrated2$scGate_multi, integrated2$seurat_clusters)
```

```{r}
tab <- table(integrated2$seurat_clusters, integrated2$functional.cluster)
tab

newlabs <- apply(tab, 1, function(x) {
  a <- which.max(x)
  names(a)})

integrated2$functional.cluster <- newlabs[integrated2$seurat_clusters]
table(integrated2$functional.cluster)

DimPlot(integrated2, reduction = "umap",
        group.by = "functional.cluster", cols=mycols, label = T) +
  ggtitle('ssSTACAS - relabeled') + theme(aspect.ratio = 1)

```

Save object to evaluate metrics
```{r}
saveRDS(integrated2, file="cache/CD8T_integrated_2nd_relabel.rds")
```

```{r fig.height=8, fig.width=5}
DefaultAssay(integrated2) <- "RNA"
Idents(integrated2) <- integrated2$functional.cluster
genes <- c('SELL','LEF1', "TCF7", "CCR7","LMNA","IL7R","GZMK","CCL4",
           "FGFBP2",'FCGR3A','XCL1',"XCL2",'GNG4',
           "CRTAM","TOX",'PDCD1',"HAVCR2","LAG3","GZMB","PRF1",
           'KLRB1','TRAV1-2','SLC4A10')
VlnPlot(integrated2, pt.size = 0.1, features = genes, cols=mycols,
        stack = TRUE, flip = TRUE, assay = "RNA", fill.by = "ident")
DefaultAssay(integrated2) <- "integrated"
```

```{r fig.height=6}
DefaultAssay(integrated2) <- "RNA"
FeaturePlot(integrated2, features=c("PDCD1","HAVCR2","TCF7","XCL2","XCL1","CD200"), ncol=3)
FeaturePlot(integrated2, features=c("TRAV1-2","SLC4A10","KLRB1","KLRC1","KLRD1","CD200"), ncol=3)
FeaturePlot(integrated2, features=c("CCR7","CRTAM","TOX","IL7R","GZMB","GZMK"), ncol=3)
FeaturePlot(integrated2, features=c("GNG4","PRF1","LAG3","CTLA4","LMNA","GZMA"), ncol=3)
DefaultAssay(integrated2) <- "integrated"
```

Saving
```{r, eval=T}
# Saving / loading
path <- paste0(path_output, "Utility.v0.4.CD8.integrated.stacas.semisup.rds")
saveRDS(integrated2, path)
```

# Integration with Harmony

```{r fig.height=6}
library(scIntegrationMetrics)

merged <- Reduce(f=merge, x=seurat.list)

merged <- FindVariableFeatures(merged, nfeatures = nfeatures) |> ScaleData() |>
  RunPCA(npcs=npcs) |> RunUMAP(reduction = "pca", dims = 1:npcs, seed.use=seed, n.neighbors = umap.neighbors)

a <- DimPlot(merged, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(merged, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a
```

```{r}
library(harmony)
set.seed(123)

harmony <- RunHarmony(merged, group.by.vars = "SampleLabel")
harmony <- RunUMAP(harmony, reduction = "harmony", dims=1:50)

saveRDS(harmony, file="cache/CD8T_harmony.rds")
```

```{r fig.height=5, width=10}
a <- DimPlot(harmony, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(harmony, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_Harmony.pdf", height=4, width=9)
```
