==> 3-Clustering_annotation.Rmd <==
---
title: "CD8 TIL Clustering + manual annotation"
author: "Paul Gueguen, Massimo Andreatta, and Santiago  Carmona"
date: "`r format(Sys.Date(),'%e de %B, %Y')`"
output: 
  rmdformats::downcute:
    lightbox: true
    thumbnails: false
    self_contained: true
    gallery: true
    code_folding: show
  pkgdown:
    as_is: true
---

```{r, include=FALSE, fig.width=16, fig.height=12}
#library(renv)
#renv::restore()
library(Seurat)
library(ggplot2)
library(scGate)
library(STACAS)
library(tidyverse)
library(SignatuR)
library(dplyr)
library(ProjecTILs)
library(RColorBrewer)
```

# Setup parameters

```{r}
myPath <- "~/Dropbox/CSI/Datasets/Human_TIL_Atlas/CD8/"

path <- "./out/Utility.v0.4.CD8.integrated.stacas.semisup.rds"
integrated <- readRDS(path)

npcs = 50
umap.min.dist = 0.3
umap.neighbors = 30
metric = "cosine"

# Set up color palette
mycols <- c('NaiveLike' = '#b7d2e0', 'CM' = '#da6f6f','EM'= '#72b28a','TEMRA' = '#e5bfaf', 'TPEX' = '#aca6e0' , 'TEX' ='#f5d39f', "MAIT" = '#fdbfd4')
mycols_scGate <- c('CD8_N' = '#b7d2e0', 'CD8_EM'= '#72b28a','CD8_TEMRA' = '#e5bfaf', 'CD8_TPEX' = '#aca6e0' , 'CD8_TEX' ='#f5d39f', "CD8_MAIT" = '#fdbfd4')

```

Rename cell types with CD8 prefix
```{r}
table(integrated$functional.cluster)

integrated$functional.cluster <- paste0("CD8.", integrated$functional.cluster)
names(mycols) <- paste0("CD8.", names(mycols))

#Reorder levels
integrated$functional.cluster <- factor(integrated$functional.cluster, levels=names(mycols))

```


# Find markers for clusters - as conserved DEG across multiple studies

```{r}
Idents(integrated) <- "functional.cluster"
DefaultAssay(integrated) <- "RNA"

set.seed(1234)

marker.list <- ProjecTILs::FindAllMarkers.bygroup(integrated, split.by = "SampleLabel",
                                 min.cells.group=30,
                                 max.cells.per.ident = 2000, 
                                 features = integrated@assays$integrated@var.features,
                                 test.use = "wilcox", logfc.threshold = 0.5,
                                 min.diff.pct=0.05, min.freq = 0.8, verbose = F)

genes <- lapply(marker.list, function(x) {
  names(x) |> head(10)
})

genes.use.bygroup <- unique(unlist(genes))
genes.use.bygroup
```


```{r fig.height=5, fig.width=7}
palettes <- list()

integrated@misc$atlas.palette <- mycols

palettes$Tissue <- colorRampPalette(brewer.pal(n = 8, 
            name = "Set1"))(length(unique(integrated$Tissue)))
names(palettes$Tissue) <- unique(integrated$Tissue)
palettes$SampleLabel <- colorRampPalette(brewer.pal(n = 8, 
            name = "Set2"))(length(unique(integrated$SampleLabel)))
names(palettes$SampleLabel) <- unique(integrated$SampleLabel)

#cluster.method <- "average"
cluster.method <- "ward.D2"
  
h <- celltype.heatmap(integrated, cluster.col = 'functional.cluster',  metadata=c("SampleLabel", "Tissue"),
                  assay = "RNA", genes = genes.use.bygroup,
                  min.cells=30, ref=integrated, cluster_genes = T, palette = palettes,
                  cluster_samples = TRUE, method=cluster.method, scale = "row",
                  brewer.palette = "RdBu", palette_reverse = T)

h

ggsave(sprintf("plots/CD8T_heatmap_ps_clustered_%s.pdf", cluster.method), plot=h, height=5, width=9)

```


To find more discriminant marker genes, focus on differences in the lower range by trimming high values

```{r}
RNA_log_cutoff <- 1
integrated@assays$RNA_trimmed <- integrated@assays$RNA
integrated@assays$RNA_trimmed@data[integrated@assays$RNA_trimmed@data > RNA_log_cutoff] <- RNA_log_cutoff
```

```{r}
Idents(integrated) <- "functional.cluster"
DefaultAssay(integrated) <- "RNA_trimmed"

sort(table(integrated$functional.cluster))
mean(table(integrated$functional.cluster,integrated$SampleLabel)>100)

set.seed(1234)

marker.list <- ProjecTILs::FindAllMarkers.bygroup(integrated, split.by = "SampleLabel",
                                 min.cells.group=30, 
                                 max.cells.per.ident = 1000, 
                                 features = NULL, test.use = "wilcox",
                                 min.diff.pct=0.05, min.freq = 2/3, verbose = F)

genes <- lapply(marker.list, function(x) {
  names(x) |> head(6)
})

genes.use.bygroup.trimm <- unique(unlist(genes))
genes.use.bygroup.trimm
```


Unsupervised heatmap with trimmed values


```{r fig.height=5, fig.width=7}
palettes <- list()

integrated@misc$atlas.palette <- mycols

palettes$Tissue <- colorRampPalette(brewer.pal(n = 8, 
            name = "Set1"))(length(unique(integrated$Tissue)))
names(palettes$Tissue) <- unique(integrated$Tissue)
palettes$SampleLabel <- colorRampPalette(brewer.pal(n = 8, 
            name = "Set2"))(length(unique(integrated$SampleLabel)))
names(palettes$SampleLabel) <- unique(integrated$SampleLabel)

cluster.method <- "average"
#cluster.method <- "ward.D2"
  
h <- celltype.heatmap(integrated, cluster.col = 'functional.cluster',  metadata=c("SampleLabel", "Tissue"),
                  assay = "RNA_trimmed", genes = genes.use.bygroup.trimm,
                  min.cells=30, ref=integrated, cluster_genes = T, palette = palettes,
                  cluster_samples = TRUE, method=cluster.method, scale = "none",
                  brewer.palette = "Blues", breaks = seq(0, RNA_log_cutoff, by = 0.01))

h

ggsave(sprintf("plots/CD8T_heatmap_pseudobulk_clustered_%s.pdf", cluster.method), plot=h, height=4, width=5)
ggsave(sprintf("plots/CD8T_heatmap_pseudobulk_clustered_%s_big.pdf", cluster.method), plot=h, height=5, width=9)



```



# Building reference from ssSTACAS integrated object

```{r}
# Transform into reference + Recompute UMAP
DefaultAssay(integrated) <- "integrated"
ref <- ProjecTILs::make.reference(integrated, ndim = npcs,
                                  recalculate.umap = T, atlas.name = "Human CD8 TILs",
                                  umap.method = "umap", min_dist = umap.min.dist,
                                  metric = metric,
                                  n_neighbors = umap.neighbors)
p <- DimPlot(ref, label=T, cols = mycols) + theme(aspect.ratio = 1)
p
ggsave("plots/CD8T_human_ref_v0.2.png", height = 4, width=6)

# Adding scGate model to the object
models <- scGate::get_scGateDB(force_update = T)

ref@misc$scGate$human <- models$human$generic$CD8TIL
ref@misc$scGate$mouse <- models$mouse$generic$CD8T

# Add color scheme to the object
ref@misc$atlas.palette <- mycols

# Refactor levels
Idents(ref) <- factor(Idents(ref), levels = names(mycols))
ref$functional.cluster <- factor(ref$functional.cluster, levels = names(mycols))

```

```{r, eval=T}
# Saving
path <- "~/Dropbox/CSI/Datasets/projectTils/CD8T_human_ref_v1_large.rds"
saveRDS(ref, file = path)

#ref <- readRDS(file = "~/Dropbox/CSI/Datasets/projectTils/CD8T_human_ref_v1_large.rds")
```


```{r}
a <- DimPlot(ref, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(ref, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_2nd.pdf", height=4, width=9)


```

```{r}
a <- DimPlot(ref, reduction = "umap",
        cols = mycols, group.by = "functional.cluster", raster = T) +
  ggtitle('Exp annotation') + theme(aspect.ratio = 1)

b <- DimPlot(ref, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_2nd_wlabels.pdf", height=4, width=9)

```

# Violin annotation

```{r fig.height=5, fig.width=14}
tmp <- ref

tmp$functional.cluster <- factor(tmp$functional.cluster, levels=rev(levels(tmp$functional.cluster)))

DefaultAssay(tmp) <- "RNA"
Idents(tmp) <- "functional.cluster"


genes <- c('SELL','LEF1',"S1PR1","IL7R", "TCF7", "CCR7",
           "LMNA","GZMA","GZMK","CCL4","FGFBP2",'FCGR3A','KLRG1',"GNLY",
           'XCL1',"XCL2","GNG4","CRTAM","TOX","CXCL13",'PDCD1',
           "CXCR6","GZMB","PRF1","CTLA4","LAG3","HAVCR2",'KLRB1','SLC4A10','TRAV1-2')

v <- VlnPlot(tmp, features = genes, stack = TRUE, flip = TRUE, fill.by = "ident", cols = mycols)

ggsave("plots/CD8_ref_subset_violins.pdf", height=12, width=6)

#Horizontal version

v <- VlnPlot(tmp, features = genes, stack = TRUE, flip = FALSE, fill.by = "ident", cols = mycols)

ggsave("plots/CD8_ref_subset_violins_horizontal.pdf", plot=v, height=4, width=12)

v
```

Dotplot visualization
```{r fig.width=12, fig.height=4}
DefaultAssay(tmp) <- "RNA"
DotPlot(tmp, features = genes, cols="RdBu", scale=T, col.max=1.5) +
  theme(axis.text.x = element_text(angle = 45, hjust=1))

ggsave("plots/CD8_ref_subset_dotplot.pdf", height=4, width=11)
```

Subset composition by study / sample
(this should be done on whole set without downsampling)

```{r}
tab <- table(ref$SampleLabel, ref$functional.cluster)
tab.norm <- tab/rowSums(tab)

df <- reshape2::melt(tab.norm)
colnames(df) <- c("Study","Subtype","Frequency")

palette <- ref@misc$atlas.palette

ggplot(df, aes(x = Study, y = Frequency, 
            fill = Subtype)) + geom_bar(stat = "identity") + 
            theme_bw() + scale_fill_manual(values = mycols) +
            theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

ggsave("plots/CD8T_subset_composition.pdf", width=7, height=4)
```

# Downsample reference

```{r eval=T}
set.seed(123)
ref.cd8.downsample <- subset(ref,  downsample = 2000)
Idents(ref.cd8.downsample) <- "functional.cluster"
DefaultAssay(ref.cd8.downsample) <- "integrated"
ref.cd8.downsample <- ProjecTILs::make.reference(ref.cd8.downsample, ndim = npcs,
                                  recalculate.umap = T, atlas.name = "Human CD8 TILs",
                                  umap.method = "umap", min_dist = umap.min.dist,
                                  metric = metric,
                                  n_neighbors = umap.neighbors)

ref.cd8.downsample$functional.cluster <- factor(ref.cd8.downsample$functional.cluster, levels = names(mycols))

#scGate models
models <- scGate::get_scGateDB(force_update = T)
ref.cd8.downsample@misc$scGate$human <- models$human$generic$CD8TIL
ref.cd8.downsample@misc$scGate$mouse <- models$mouse$generic$CD8T


# Add color scheme to the object
ref.cd8.downsample@misc$atlas.palette <- mycols

p <- DimPlot(ref.cd8.downsample, label=T, cols = mycols) + theme(aspect.ratio = 1)
p

```

```{r, eval=T}

path <- "~/Dropbox/CSI/Datasets/projectTils/CD8T_human_ref_v1.rds"
saveRDS(ref.cd8.downsample, file = path)

#ref.cd8.downsamples <- readRDS(file = "~/Dropbox/CSI/Datasets/projectTils/CD8T_human_ref_v1.rds")
```

```{r}
a <- DimPlot(ref.cd8.downsample, reduction = "umap",
        cols = mycols_scGate, group.by = "scGate_multi", raster = T) +
  ggtitle('scGate_multi') + theme(aspect.ratio = 1)

b <- DimPlot(ref.cd8.downsample, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_downsample.pdf", height=4, width=9)


```

```{r}
a <- DimPlot(ref.cd8.downsample, reduction = "umap",
        cols = mycols, group.by = "functional.cluster", raster = T) +
  ggtitle('Exp annotation') + theme(aspect.ratio = 1)

b <- DimPlot(ref.cd8.downsample, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_ssIntegration_downsample_wlabels.pdf", height=4, width=9)

```

```{r}
DimPlot(ref.cd8.downsample, reduction = "umap",
        cols = ref.cd8.downsample@misc$atlas.palette,
        group.by = "functional.cluster", label = T) + 
  theme(aspect.ratio = 1) + NoLegend()
```
 
# Violin annotation

```{r fig.height=10, fig.width=5}
DefaultAssay(ref.cd8.downsample) <- "RNA"
Idents(ref.cd8.downsample) <- "functional.cluster"


VlnPlot(ref.cd8.downsample, features = genes, stack = TRUE, flip = TRUE, fill.by = "ident", cols = mycols)
DefaultAssay(ref.cd8.downsample) <- "integrated"

ggsave("plots/CD8_ref_subset_violins_ds.pdf", height=7, width=6)
```
