==> 4-Metrics.Rmd <==
---
title: "Integrate DC datasets"
author: "Massimo Andreatta & Santiago Carmona"
#output: html_notebook
---

```{r}
renv::restore()
```

Libraries
```{r}
#remotes::install_github('carmonalab/STACAS')
#remotes::install_github('carmonalab/scIntegrationMetrics')

library(ggplot2)
library(Seurat)
library(tidyr)
library(dplyr)
library(RColorBrewer)
library(SeuratObject)
library(Matrix)
library(patchwork)
library(remotes)
library(STACAS)
library(scIntegrationMetrics)
```

Read in objects for uncorrected, unsupervised and semi-sup integration
```{r}
merged <- readRDS("cache/CD8T_unintegrated.rds")
integrated.unsup <- readRDS("cache/CD8T_integrated_unsup.rds")
integrated.ss <- readRDS("cache/CD8T_integrated_1st.rds")
integrated.ss.2nd <- readRDS("cache/CD8T_integrated_2nd.rds")
harmony <- readRDS("cache/CD8T_harmony.rds")

annotated <- readRDS("~/Dropbox/CSI/Datasets/projectTils/CD8T_human_ref_v0.2.rds")
```

Convert scANVI object
```{r}
h5ad_file = "out/scANVI_integrated_CD8.h5ad"

Convert(h5ad_file, dest = "h5seurat", assay="RNA", overwrite = TRUE)
scanvi <- LoadH5Seurat("out/scANVI_integrated_CD8.h5seurat", misc=FALSE, meta.data=FALSE)

merged.meta <- merged@meta.data[,c("SampleLabel","scGate_multi")]
scanvi <- AddMetaData(scanvi, merged.meta)
scanvi@reductions$pca <- scanvi@reductions$emb
```


```{r}
# Setup color palette
mycols <- c('NaiveLike' = '#b7d2e0', 'CM' = '#da6f6f','EM'= '#72b28a','TEMRA' = '#e5bfaf', 'TPEX' = '#aca6e0' , 'TEX' ='#f5d39f', "MAIT" = '#fdbfd4')

mycols_scGate <- c('CD8_N' = '#b7d2e0', 'CD8_EM'= '#72b28a','CD8_TEMRA' = '#e5bfaf', 'CD8_TPEX' = '#aca6e0' , 'CD8_TEX' ='#f5d39f', "CD8_MAIT" = '#fdbfd4')
```

```{r fig.height=7, fig.width=12}
a <- DimPlot(merged, group.by = "scGate_multi", cols = mycols_scGate) + 
  ggtitle("Unintegrated") + NoLegend() + theme(aspect.ratio = 1)
b <- DimPlot(integrated.unsup, group.by = "scGate_multi", cols = mycols_scGate) + 
  ggtitle("STACAS") + NoLegend() + theme(aspect.ratio = 1)
c <- DimPlot(integrated.ss, group.by = "scGate_multi", cols = mycols_scGate) +
  ggtitle("ssSTACAS (1)") + theme(aspect.ratio = 1)
d <- DimPlot(integrated.ss.2nd, group.by = "scGate_multi", cols = mycols_scGate) + 
  ggtitle("ssSTACAS (2)") + NoLegend() + theme(aspect.ratio = 1)
e <- DimPlot(harmony, group.by = "scGate_multi", cols = mycols_scGate) +
  ggtitle("Harmony") + NoLegend() + theme(aspect.ratio = 1)
f <- DimPlot(scanvi, group.by = "scGate_multi", cols = mycols_scGate) +
  ggtitle("scANVI") + theme(aspect.ratio = 1)

(a | b | c) / (d | e | f)
```

Subset on the same cells to calculate metrics (some cells were removed during integration)
```{r}
cells.merged <- colnames(merged)
cells.us <- colnames(integrated.unsup)
cells.i1 <- colnames(integrated.ss)
cells.i2 <- colnames(integrated.ss.2nd)
cells.ha <- colnames(harmony)
cells.sc <- colnames(scanvi)

cells.keep <- intersect(cells.merged, cells.i2)

#Subset on common cells
merged <- subset(merged, cells=cells.keep)
integrated.unsup <- subset(integrated.unsup, cells=cells.keep)
integrated.ss <- subset(integrated.ss, cells=cells.keep)
integrated.ss.2nd <- subset(integrated.ss.2nd, cells=cells.keep)
harmony <- subset(harmony, cells=cells.keep)
scanvi <- subset(scanvi, cells=cells.keep)
```

Calculate integration metrics based on scGate labels (prior knowledge from standard markers)
```{r}
#Integration metrics
integrationMetrics <- list()

metricsLabels <- NULL  #use all labels
meta.label <- "scGate_multi"
meta.batch <- "SampleLabel"

integrationMetrics[["Uncorrected"]] <- getIntegrationMetrics(object=merged,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      metricsLabels = metricsLabels)

integrationMetrics[["STACAS"]] <- getIntegrationMetrics(object=integrated.unsup,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      metricsLabels = metricsLabels)

integrationMetrics[["ssSTACAS(1)"]] <- getIntegrationMetrics(object=integrated.ss,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      metricsLabels = metricsLabels)

integrationMetrics[["ssSTACAS(2)"]] <- getIntegrationMetrics(object=integrated.ss.2nd,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      metricsLabels = metricsLabels)

integrationMetrics[["Harmony"]] <- getIntegrationMetrics(object=harmony,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      method.reduction = "harmony",
                                                      metricsLabels = metricsLabels)

integrationMetrics[["scanvi"]] <- getIntegrationMetrics(object=scanvi,
                                                      meta.label = meta.label,
                                                      meta.batch = meta.batch,
                                                      metricsLabels = metricsLabels)

```

Combine integration metrics
```{r fig.height=7}
integrationMetricsSummary <- data.frame(unlist(integrationMetrics)) %>%
  tibble::rownames_to_column() %>%
  dplyr::rename(value=unlist.integrationMetrics.) %>%
  tidyr::separate(rowname, c("Method","Metric"), sep="\\.")


integrationMetricsSummary$Method <- factor(integrationMetricsSummary$Method,
                                           levels=c("Uncorrected","STACAS",
                                                    "ssSTACAS(1)","ssSTACAS(2)","Harmony","scanvi"))

ggplot(integrationMetricsSummary, aes(x=Method, y=value, fill=Method)) + geom_bar(stat="identity") + 
    theme_bw()  +
    theme(legend.position="none",
          axis.text.x=element_text(angle = 45, vjust = 1, hjust=1)) +
    xlab("Method") + facet_wrap(~Metric, scales = "free") +
    scale_fill_brewer(palette="Set1")
```


Scatterplot of two performance metrics
```{r fig.width=8, fig.height=4}
library(RColorBrewer)
library(tidyverse)
palette <- brewer.pal(n=length(unique(integrationMetricsSummary$Method)), name = "Set1")

a <- integrationMetricsSummary %>% dplyr::filter(Metric %in% c("celltype_ASW","CiLISI")) %>%
  pivot_wider(names_from = Metric, values_from = value ) %>%
  ggplot(aes(x=CiLISI, y=`celltype_ASW`, label=Method)) +
  geom_point(aes(color=Method), size=2) + geom_text(hjust=0.1, vjust=0) + 
  scale_color_manual(values = palette) + theme_light()

a

ggsave("plots/CD8T_metrics_scatterplot.pdf", height=3, width=5)
```

Use final annotations

```{r}
labels <- annotated$functional.cluster

# Set up color palette
mycols <- c('CD8.NaiveLike' = '#b7d2e0', 'CD8.CM' = '#da6f6f',
            'CD8.EM'= '#72b28a','CD8.TEMRA' = '#e5bfaf', 
            'CD8.TPEX' = '#aca6e0' , 'CD8.TEX' ='#f5d39f', "CD8.MAIT" = '#fdbfd4')

merged <- subset(merged, cells=Cells(annotated))
merged@meta.data[names(labels),"expert.labels"] <- labels

a <- DimPlot(merged, reduction = "umap",
        cols = mycols, group.by = "expert.labels", raster = T) +
  ggtitle('Expert annot') + theme(aspect.ratio = 1)

b <- DimPlot(merged, reduction = "umap",
        group.by = "SampleLabel", raster = T) +
  theme(aspect.ratio = 1, legend.text = element_text(size=7),
        legend.key.size = unit(0.1, 'cm'))

b | a

ggsave("plots/umap_preintegration_byExpLabels.pdf", height=4, width=9)

```
